implemented new filter reprentation
