+++ src/Dashboard/Filter.php
@@ -34,19 +34,18 @@
  */
 
 namespace Glpi\Dashboard;
-
 use Glpi\Plugin\Hooks;
-use Group;
-use Html;
-use ITILCategory;
-use Location;
-use Manufacturer;
-use Plugin;
-use RequestType;
 use Session;
-use State;
-use Ticket;
-use User;
+use Glpi\Dashboard\Filters\DatesFilter;
+use Glpi\Dashboard\Filters\ItilCategoryFilter;
+use Glpi\Dashboard\Filters\LocationFilter;
+use Glpi\Dashboard\Filters\ManufacturerFilter;
+use Glpi\Dashboard\Filters\RequestTypeFilter;
+use Glpi\Dashboard\Filters\StateFilter;
+use Glpi\Dashboard\Filters\TicketTypeFilter;
+use Glpi\Dashboard\Filters\GroupTechFilter;
+use Glpi\Dashboard\Filters\UserTechFilter;
+use Glpi\Dashboard\Filters\DatesModFilter;
 
 /**
  * Filter class
@@ -58,234 +57,38 @@ class Filter extends \CommonDBChild
 
     /**
      * Return all available filters
-     * Plugins can hooks on this functions to add their own filters
      *
      * @return array of filters
      */
     public static function getAll(): array
     {
+        global $PLUGIN_HOOKS;
+        $more_filters = [];
+
         $filters = [
-            'dates'        => __("Creation date"),
-            'dates_mod'    => __("Last update"),
-            'itilcategory' => ITILCategory::getTypeName(Session::getPluralNumber()),
-            'requesttype'  => RequestType::getTypeName(Session::getPluralNumber()),
-            'location'     => Location::getTypeName(Session::getPluralNumber()),
-            'manufacturer' => Manufacturer::getTypeName(Session::getPluralNumber()),
-            'state'        => State::getTypeName(Session::getPluralNumber()),
-            'tickettype'   => _n("Ticket type", "Ticket types", Session::getPluralNumber()),
-            'group_tech'   => __("Technician group"),
-            'user_tech'    => __("Technician"),
+            DatesFilter::class,
+            DatesModFilter::class,
+            ItilCategoryFilter::class,
+            LocationFilter::class,
+            ManufacturerFilter::class,
+            RequestTypeFilter::class,
+            StateFilter::class,
+            TicketTypeFilter::class,
+            GroupTechFilter::class,
+            UserTechFilter::class,
         ];
 
-        $more_filters = Plugin::doHookFunction(Hooks::DASHBOARD_FILTERS);
-        if (is_array($more_filters)) {
-            $filters = array_merge($filters, $more_filters);
+        if(isset($PLUGIN_HOOKS[Hooks::DASHBOARD_FILTERS])) {
+            foreach ($PLUGIN_HOOKS[Hooks::DASHBOARD_FILTERS] as $hook_filters) {
+                foreach ($hook_filters as $filter) {
+                    $more_filters[] = $filter;
+                }
+            }
         }
+        $filters = array_merge($filters, $more_filters);
 
         return $filters;
-    }
-
-    /**
-     * Get HTML for a dates range filter
-     *
-     * @param string|array $values init the input with these values, will be a string if empty values
-     * @param string $fieldname how is named the current date field
-     *                         (used to specify creation date or last update)
-     *
-     * @return string
-     */
-    public static function dates($values = "", string $fieldname = 'dates'): string
-    {
-       // string mean empty value
-        if (is_string($values)) {
-            $values = [];
-        }
-
-        $rand  = mt_rand();
-        $label = self::getAll()[$fieldname];
-        $field = Html::showDateField('filter-dates', [
-            'value'        => $values,
-            'rand'         => $rand,
-            'range'        => true,
-            'display'      => false,
-            'calendar_btn' => false,
-            'placeholder'  => $label,
-            'on_change'    => "on_change_{$rand}(selectedDates, dateStr, instance)",
-        ]);
-
-        $js = <<<JAVASCRIPT
-      var on_change_{$rand} = function(selectedDates, dateStr, instance) {
-         // we are waiting for empty value or a range of dates,
-         // don't trigger when only the first date is selected
-         var nb_dates = selectedDates.length;
-         if (nb_dates == 0 || nb_dates == 2) {
-            Dashboard.getActiveDashboard().saveFilter('{$fieldname}', selectedDates);
-            $(instance.input).closest("fieldset").addClass("filled");
-         }
-      };
-JAVASCRIPT;
-        $field .= Html::scriptBlock($js);
-
-        return self::field($fieldname, $field, $label, is_array($values) && count($values) > 0);
-    }
-
-    /**
-     * Get HTML for a dates range filter. Same as date but for last update field
-     *
-     * @param string|array $values init the input with these values, will be a string if empty values
-     *
-     * @return string
-     */
-    public static function dates_mod($values): string
-    {
-        return self::dates($values, "dates_mod");
-    }
-
-
-    public static function itilcategory(string $value = ""): string
-    {
-        return self::displayList($value, 'itilcategory', ITILCategory::class);
-    }
-
-    public static function requesttype(string $value = ""): string
-    {
-        return self::displayList($value, 'requesttype', RequestType::class);
-    }
-
-    public static function location(string $value = ""): string
-    {
-        return self::displayList($value, 'location', Location::class);
-    }
-
-    public static function manufacturer(string $value = ""): string
-    {
-        return self::displayList($value, 'manufacturer', Manufacturer::class);
-    }
-
-    public static function group_tech(string $value = ""): string
-    {
-        return self::displayList($value, 'group_tech', Group::class, ['toadd' => ['mygroups' => __("My groups")]]);
-    }
-
-    public static function user_tech(string $value = ""): string
-    {
-        return self::displayList($value, 'user_tech', User::class, [
-            'right' => 'own_ticket',
-            'toadd' => [
-                [
-                    'id'    => 'myself',
-                    'text'  => __('Myself'),
-                ]
-            ]
-        ]);
-    }
-
-    public static function state(string $value = ""): string
-    {
-        return self::displayList($value, 'state', State::class);
-    }
-
-    public static function tickettype(string $value = ""): string
-    {
-        return self::displayList($value, 'tickettype', Ticket::class, [
-            'condition' => ['id' => -1],
-            'toadd'     => Ticket::getTypes()
-        ]);
-    }
-
-    public static function displayList(
-        string $value = "",
-        string $fieldname = "",
-        string $itemtype = "",
-        array $add_params = []
-    ): string {
-        $value     = !empty($value) ? $value : null;
-        $rand      = mt_rand();
-        $label     = self::getAll()[$fieldname];
-        $field     = $itemtype::dropdown([
-            'name'                => $fieldname,
-            'value'               => $value,
-            'rand'                => $rand,
-            'display'             => false,
-            'display_emptychoice' => false,
-            'emptylabel'          => '',
-            'placeholder'         => $label,
-            'on_change'           => "on_change_{$rand}()",
-            'allowClear'          => true,
-            'width'               => ''
-        ] + $add_params);
-
-        $js = <<<JAVASCRIPT
-      var on_change_{$rand} = function() {
-         var dom_elem    = $('#dropdown_{$fieldname}{$rand}');
-         var selected    = dom_elem.find(':selected').val();
-
-         Dashboard.getActiveDashboard().saveFilter('{$fieldname}', selected);
-
-         $(dom_elem).closest("fieldset").toggleClass("filled", selected !== null)
-      };
-
-JAVASCRIPT;
-        $field .= Html::scriptBlock($js);
-
-        return self::field($fieldname, $field, $label, $value !== null);
-    }
-
-    /**
-     * Get generic HTML for a filter
-     *
-     * @param string $id system name of the filter (ex "dates")
-     * @param string $field html of the filter
-     * @param string $label displayed label for the filter
-     * @param bool   $filled
-     *
-     * @return string the html for the complete field
-     */
-    public static function field(
-        string $id = "",
-        string $field = "",
-        string $label = "",
-        bool $filled = false
-    ): string {
-
-        $rand  = mt_rand();
-        $class = $filled ? "filled" : "";
-
-        $js = <<<JAVASCRIPT
-      $(function () {
-         $('#filter-{$rand} input')
-            .on('input', function() {
-               var str_len = $(this).val().length;
-               if (str_len > 0) {
-                  $('#filter-{$rand}').addClass('filled');
-               } else {
-                  $('#filter-{$rand}').removeClass('filled');
-               }
-
-               $(this).width((str_len + 1) * 8 );
-            });
-
-         $('#filter-{$rand}')
-            .hover(function() {
-               $('.dashboard .card.filter-{$id}').addClass('filter-impacted');
-            }, function() {
-               $('.dashboard .card.filter-{$id}').removeClass('filter-impacted');
-            });
-      });
-JAVASCRIPT;
-        $js = Html::scriptBlock($js);
-
-        $html  = <<<HTML
-      <fieldset id='filter-{$rand}' class='filter $class' data-filter-id='{$id}'>
-         $field
-         <legend>$label</legend>
-         <i class='btn btn-sm btn-icon btn-ghost-secondary ti ti-trash delete-filter'></i>
-         {$js}
-      </fieldset>
-HTML;
-
-        return $html;
-    }
+    }   
 
     /**
      * Return filters for the provided dashboard
@@ -297,6 +100,7 @@ HTML;
     public static function getForDashboard(int $dashboards_id = 0): string
     {
         global $DB;
+
 
         $dr_iterator = $DB->request([
             'FROM'  => self::getTable(),
