+++ src/Dashboard/Filters/DatesModFilter.php
@@ -1,0 +1,76 @@
+<?php
+
+/**
+ * ---------------------------------------------------------------------
+ *
+ * GLPI - Gestionnaire Libre de Parc Informatique
+ *
+ * http://glpi-project.org
+ *
+ * @copyright 2015-2023 Teclib' and contributors.
+ * @copyright 2003-2014 by the INDEPNET Development Team.
+ * @licence   https://www.gnu.org/licenses/gpl-3.0.html
+ *
+ * ---------------------------------------------------------------------
+ *
+ * LICENSE
+ *
+ * This file is part of GLPI.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * ---------------------------------------------------------------------
+ */
+
+namespace Glpi\Dashboard\Filters;
+
+use Glpi\Dashboard\Filters\AbstractFilter;
+use Glpi\Dashboard\Filters\DatesFilter;
+use Html;
+
+
+class DatesModFilter extends AbstractFilter
+{
+    /**
+     * Get the filter name
+     *
+     * @return string
+     */
+    public static function getName(): string
+    {
+        return __("Last update");
+    }
+
+    /**
+     * Get the filter id
+     *
+     * @return string
+     */
+    public static function getId() : string
+    {
+        return "dates_mod";
+    }
+
+    /**
+     * Get HTML for a dates range filter. Same as date but for last update field
+     *
+     * @param string|array $values init the input with these values, will be a string if empty values
+     *
+     * @return string
+     */
+    public static function getHtml($values = ""): string
+    {
+        return DatesFilter::getHtml($values, "dates_mod");
+    }
+}
